Rename directory name for shared styles
